+++ frontend/src/redux/auth/action-types.ts
@@ -1,5 +1,5 @@
-import { Auth, ReduxAction } from "../../types"
+import { AuthenticationState, Action } from "../../types"
 
 export const UPDATE_AUTH = "UPDATE_AUTH"
 export type UpdateAuth = typeof UPDATE_AUTH
-export interface UpdateAuthAction extends ReduxAction<UpdateAuth, Auth> { }
+export interface UpdateAuthAction extends Action<UpdateAuth, AuthenticationState> { }
