+++ frontend/src/redux/auth/index.ts
@@ -1,2 +1,1 @@
-
 export { authReducer } from "./reducer"
